Add option to remove suspect very high damages
In at least one lising, it looked like one event had the state-level total for damages, rather than a county-specific estimate (Floyd, North Carolina). If the user chooses, now `find_events` can look for any damage listings where a single event listing exceeds the total for the damages for all other events in the state combined. Default is to not make this adjustment.